+++ dashboard/controller/users/editUser.js
@@ -22,14 +22,14 @@ module.exports = function editUser(req, res) {
 			req.body.password = req.body.password.trim();
 			req.body.color = JSON.parse(req.body.color);
 			req.assert('name', common.l10n.get('UsernameInvalid')).matches(regexValidate("user"));
-			req.assert('password', common.l10n.get('PasswordAtLeast', {count:users.ini().security.min_password_size})).len(users.ini().security.min_password_size);
+			req.assert('password', common.l10n.get('PasswordAtLeast', { count: users.ini().security.min_password_size })).len(users.ini().security.min_password_size);
 			req.assert('password', common.l10n.get('PasswordInvalid')).matches(regexValidate("pass"));
 			req.body.classrooms = req.body.classrooms || [];
 			if (typeof req.body.classrooms == 'string') {
 				req.body.classrooms = [req.body.classrooms];
 			}
 
-			if (req.session &&  req.session.user && req.session.user.user && req.session.user.user.role == "teacher") {
+			if (req.session && req.session.user && req.session.user.user && req.session.user.user.role == "teacher") {
 				if ((!req.body.classrooms) || !(req.body.classrooms.length > 0)) {
 					req.flash('errors', {
 						msg: common.l10n.get('EmptyClassroom')
@@ -50,12 +50,12 @@ module.exports = function editUser(req, res) {
 					})
 					.end(function (error, response) {
 						if (response.statusCode == 200) {
-							if (response.body.role=="student" && typeof(req.body.classrooms) == "object") {
-								dashboard_utils.getAllClassrooms(req, res, function(classrooms) {
+							if (response.body.role == "student" && typeof (req.body.classrooms) == "object") {
+								dashboard_utils.getAllClassrooms(req, res, function (classrooms) {
 									if (classrooms && classrooms.classrooms && classrooms.classrooms.length > 0) {
 										var classroomCounter = 0;
 										var classroomRequests = 0;
-										for (var i=0; i<classrooms.classrooms.length; i++) {
+										for (var i = 0; i < classrooms.classrooms.length; i++) {
 											classrooms.classrooms[i].students = classrooms.classrooms[i].students || [];
 											if (req.body.classrooms.includes(classrooms.classrooms[i]._id) || classrooms.classrooms[i].students.includes(response.body._id)) {
 												classroomRequests++;
@@ -68,7 +68,7 @@ module.exports = function editUser(req, res) {
 														classrooms.classrooms[i].students = [response.body._id];
 													}
 												} else {
-													classrooms.classrooms[i].students = classrooms.classrooms[i].students.filter(function(value){
+													classrooms.classrooms[i].students = classrooms.classrooms[i].students.filter(function (value) {
 														return (value != response.body._id && value != null);
 													});
 												}
@@ -85,7 +85,7 @@ module.exports = function editUser(req, res) {
 														classroomCounter++;
 														if (classroomCounter == classroomRequests) {
 															req.flash('success', {
-																msg: common.l10n.get('UserUpdated', {name: req.body.name})
+																msg: common.l10n.get('UserUpdated', { name: req.body.name })
 															});
 															return res.redirect('/dashboard/users/');
 														}
@@ -95,21 +95,21 @@ module.exports = function editUser(req, res) {
 										if (classroomRequests == 0) {
 											if (classroomCounter == classroomRequests) {
 												req.flash('success', {
-													msg: common.l10n.get('UserUpdated', {name: req.body.name})
+													msg: common.l10n.get('UserUpdated', { name: req.body.name })
 												});
 												return res.redirect('/dashboard/users/');
 											}
 										}
 									} else {
 										req.flash('success', {
-											msg: common.l10n.get('UserUpdated', {name: req.body.name})
+											msg: common.l10n.get('UserUpdated', { name: req.body.name })
 										});
 										return res.redirect('/dashboard/users/');
 									}
 								});
 							} else {
 								req.flash('success', {
-									msg: common.l10n.get('UserUpdated', {name: req.body.name})
+									msg: common.l10n.get('UserUpdated', { name: req.body.name })
 								});
 								if (response.body.role == "admin") {
 									// send to admin page
@@ -124,10 +124,10 @@ module.exports = function editUser(req, res) {
 							}
 						} else {
 							req.flash('errors', {
-								msg: common.l10n.get('ErrorCode'+response.body.code)
+								msg: common.l10n.get('ErrorCode' + response.body.code)
 							});
 							return res.redirect('/dashboard/users/edit/' + req.params.uid);
-						}	
+						}
 					});
 			} else {
 				req.flash('errors', errors);
@@ -147,9 +147,9 @@ module.exports = function editUser(req, res) {
 					} else if (response.statusCode == 200) {
 						if (user && user.role == 'teacher') {
 							// fetch classrooms
-							dashboard_utils.getAllClassrooms(req, res, function(classrooms) {
-								if (user.classrooms && typeof(user.classrooms) == "object" && user.classrooms.length > 0 && classrooms && classrooms.classrooms && classrooms.classrooms.length > 0) {
-									for (var i=0; i<classrooms.classrooms.length; i++) {
+							dashboard_utils.getAllClassrooms(req, res, function (classrooms) {
+								if (user.classrooms && typeof (user.classrooms) == "object" && user.classrooms.length > 0 && classrooms && classrooms.classrooms && classrooms.classrooms.length > 0) {
+									for (var i = 0; i < classrooms.classrooms.length; i++) {
 										if (user.classrooms.indexOf(classrooms.classrooms[i]._id) != -1) {
 											classrooms.classrooms[i]['is_member'] = true;
 										}
@@ -168,9 +168,9 @@ module.exports = function editUser(req, res) {
 								});
 							});
 						} else if (user && user.role == 'student') {
-							dashboard_utils.getAllClassrooms(req, res, function(classrooms) {
+							dashboard_utils.getAllClassrooms(req, res, function (classrooms) {
 								if (classrooms && classrooms.classrooms && classrooms.classrooms.length > 0) {
-									for (var i=0; i<classrooms.classrooms.length; i++) {
+									for (var i = 0; i < classrooms.classrooms.length; i++) {
 										if (typeof classrooms.classrooms[i].students == "object" && classrooms.classrooms[i].students.includes(user._id)) {
 											classrooms.classrooms[i]['is_member'] = true;
 										}
@@ -203,10 +203,10 @@ module.exports = function editUser(req, res) {
 						}
 					} else {
 						req.flash('errors', {
-							msg: common.l10n.get('ErrorCode'+user.code)
+							msg: common.l10n.get('ErrorCode' + user.code)
 						});
 						return res.redirect('/dashboard/users');
-					} 
+					}
 				});
 		}
 	} else {